analysis/requirements.txt
albumentations==1.2.1
appdirs==1.4.4
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attr==0.3.1
attrs==21.4.0
azure-core==1.24.2
azure-storage-blob==12.13.0
backcall==0.2.0
boto==2.49.0
boto3==1.16.63
botocore==1.19.63
Bottleneck==1.3.5
boxing==0.1.4
brotlipy==0.7.0
cachetools==4.2.4
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.1.14
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rest-swagger==2.2.0
django-rq==2.5.1
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-yasg==1.20.0
entrypoints==0.4
executing==0.8.3
expiringdict==1.1.4
fonttools==4.34.4
google-api-core==1.31.5
google-auth==1.35.0
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==1.5.0
google-cloud-logging==2.7.2
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.47.0
htmlmin==0.1.12
idna==3.3
imageio==2.20.0
inflection==0.5.1
ipykernel==6.15.0
ipython==8.4.0
isodate==0.6.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.3.4
jupyter-core==4.10.0
kiwisolver==1.4.4
label-studio==1.5.0.post0
label-studio-converter==0.0.40
label-studio-tools==0.0.0.dev14
launchdarkly-server-sdk==7.3.0
lockfile==0.12.2
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
msrest==0.7.1
nest-asyncio==1.5.5
networkx==2.8.5
nltk==3.6.7
numexpr==2.8.3
numpy==1.23.1
oauthlib==3.2.0
openapi-codec==1.3.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
prompt-toolkit==3.0.29
proto-plus==1.20.6
protobuf==3.20.1
psutil==5.9.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
pydantic==1.8.2
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
qudida==0.0.4
redis==4.3.4
regex==2022.7.9
requests==2.28.1
requests-oauthlib==1.3.1
rq==1.10.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
rules==2.2
s3transfer==0.3.7
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
semver==2.13.0
sentry-sdk==1.7.0
setuptools==61.2.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
threadpoolctl==3.1.0
tifffile==2022.5.4
torch==1.12.0
torchaudio==0.12.0
torchmetrics==0.9.3
torchvision==0.13.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
ua-parser==0.15.0
ujson==5.4.0
uritemplate==4.1.1
urllib3==1.26.10
user-agents==2.2.0
wcwidth==0.2.5
wheel==0.37.1
wrapt==1.14.1
xmljson==0.2.0
